requirements.txt
pymongo==4.5.0
pinecone-client>=2.2.0
setuptools>=44.0.0
# Install huggingface_hub before sentence-transformers to ensure compatibility
huggingface_hub==0.12.0
sentence-transformers>=2.2.0
python-dotenv>=0.21.0
transformers
torch>=1.6.0
uuid>=1.30
